Adds router functionality
