+++ package.json
@@ -13,6 +13,7 @@
     "html-react-parser": "^5.1.10",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.22.3",
     "react-scripts": "5.0.1",
     "style-loader": "^3.3.4",
     "styled-components": "^6.1.8",
