+++ src/components/Error/Error.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components'
+
+function Error() {
+
+
+    return (
+        <>
+            <div>Error</div>
+        </>
+    )
+
+
+}
+
+export {Error}
