+++ src/components/LandingPage/LandingPage.js
@@ -1,0 +1,16 @@
+import styled from 'styled-components'
+import { Link } from 'react-router-dom'
+
+function LandingPage() {
+
+
+
+    return (
+        <>
+            <Link to='/HomePage'>**WikiLinks**</Link>
+        </>
+    )
+
+}
+
+export {LandingPage}
