+++ src/components/Stats/Stats.js
@@ -1,0 +1,16 @@
+import styled from 'styled-components'
+
+
+function Stats() {
+
+
+    return (
+        <>
+            <h1>StatsGoHere</h1>
+        </>
+    )
+}
+
+
+export { Stats }
+
